Move interval index tests to indexes/test_interval.py
